Isolate TurboSnap version 1.0
