+++ node-src/git/git.test.ts
@@ -5,7 +5,7 @@ import {
   AncestorMissingError,
   BaselineCheckoutFailedError,
   GitCommandError,
-} from '../lib/turbosnap/errors';
+} from '../lib/turbosnap/1.0/errors';
 import * as execGit from './execGit';
 import {
   checkoutFile,
