+++ node-src/lib/turbosnap/1.0/findChangedPackageFiles.ts
@@ -1,6 +1,6 @@
-import { execGitCommand } from '../../git/execGit';
-import { Context } from '../../types';
-import { isPackageMetadataFile } from '../utilities';
+import { execGitCommand } from '../../../git/execGit';
+import { Context } from '../../../types';
+import { isPackageMetadataFile } from '../../utilities';
 
 // TODO: refactor this function
 // eslint-disable-next-line complexity
